data_service/pyproject.toml
[project]
name = "data-service"
version = "0.1.0"
description = "Data service with hardcoded responses (no database dependency)"
requires-python = ">=3.9"
dependencies = [
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.scripts]
data-service = "data_service.server:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
